Move TAlignedDeviceHandler to separate cpp
